requirements/testing.txt
-r base.txt

pytest
pytest-django
factory-boy
